[#4] Refactor: 패키지 구조 변경
- controller, service, repository로 구성되어있던 패키지 구조를 도메인별 패키지 구조로 변경
