+++ BE/src/main/java/louie/dong/airbnb/banner/BannerController.java
@@ -1,4 +1,4 @@
-package louie.dong.airbnb.controller;
+package louie.dong.airbnb.banner;
 
 import java.util.Collections;
 import java.util.Map;
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-public class MainController {
+public class BannerController {
 
     private static final String MAIN_IMAGE_URL = "https://user-images.githubusercontent.com/92966772/169965428-e12f898a-1f72-4510-8ddb-8086519366c0.PNG";
