+++ BE/src/main/java/louie/dong/airbnb/travel/TravelController.java
@@ -1,19 +1,18 @@
-package louie.dong.airbnb.controller;
+package louie.dong.airbnb.travel;
 
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import louie.dong.airbnb.dto.PopularDestinationResponse;
-import louie.dong.airbnb.service.MockPopularDestinationService;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
-public class PopularDestinationController {
+public class TravelController {
 
-    private final MockPopularDestinationService mockPopularDestinationService;
+    private final MockTravelService mockPopularDestinationService;
 
-    @GetMapping("/popular-destinations")
+    @GetMapping("/travel/popular")
     public List<PopularDestinationResponse> getPopularDestinations() {
         return mockPopularDestinationService.findAll();
     }
